adding dependencies renaming packages
