+++ src/main/resources/junit-platform.properties
@@ -1,0 +1,1 @@
+junit.jupiter.displayname.generator.default=io.github.timofeevvr.petstore.TestNameGenerator
