+++ src/test/java/io/github/timofeevvr/petstore/StoreApiIntegrationTest.java
@@ -1,7 +1,7 @@
-package com.petstore.integration.tests;
+package io.github.timofeevvr.petstore;
 
-import com.petstore.integration.tests.model.Order;
-import com.petstore.integration.tests.model.Pet;
+import io.github.timofeevvr.petstore.model.Order;
+import io.github.timofeevvr.petstore.model.Pet;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.TestInstance;
@@ -10,26 +10,19 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import static com.petstore.integration.tests.model.Order.StatusEnum.PLACED;
-import static com.petstore.integration.tests.model.Pet.StatusEnum.AVAILABLE;
 import static java.net.HttpURLConnection.HTTP_OK;
 import static org.hamcrest.Matchers.*;
 
 @Slf4j
 @Tag("store")
-@SpringBootTest
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-class StoreApiIntegrationTest {
-
-    @Autowired
-    private ApiClient petstoreClient;
+class StoreApiIntegrationTest extends TestBase {
 
     @ParameterizedTest
     @ValueSource(longs = {1, 2, 5, 10})
     void givenExistingPet_whenPlaceOrderForThePet_thenOrderStatusIsPlaced(long orderId) {
         log.info("WHEN. FIND AVAILABLE PET ID");
         Long existingPetId = petstoreClient.pet().findPetsByStatus()
-                .statusQuery(AVAILABLE)
+                .statusQuery(Pet.StatusEnum.AVAILABLE)
                 .execute(response -> response)
                 .then()
                 .log()
@@ -45,7 +38,7 @@ class StoreApiIntegrationTest {
                 .id(orderId)
                 .petId(existingPetId)
                 .quantity(1)
-                .status(PLACED);
+                .status(Order.StatusEnum.PLACED);
         log.info("WHEN. PLACE ORDER");
         petstoreClient.store().placeOrder()
                 .body(order)
@@ -69,6 +62,6 @@ class StoreApiIntegrationTest {
                 .body()
                 .assertThat()
                 .statusCode(HTTP_OK)
-                .body("status", equalTo(PLACED.getValue()));
+                .body("status", equalTo(Order.StatusEnum.PLACED.getValue()));
     }
 }
