--- src/test/resources/junit-platform.properties
@@ -1,1 +1,0 @@
-junit.jupiter.displayname.generator.default=com.petstore.integration.tests.TestNameGenerator
